refactor: 기존의 회원가입 로직 관련 클래스들 이름 수정
