+++ src/main/java/io/wisoft/avocadobackendhexagonal/domain/auth/adapter/in/web/SignupController.java
@@ -1,8 +1,8 @@
 package io.wisoft.avocadobackendhexagonal.domain.auth.adapter.in.web;
 
-import io.wisoft.avocadobackendhexagonal.domain.auth.adapter.in.web.dto.SignupRequest;
+import io.wisoft.avocadobackendhexagonal.domain.auth.adapter.in.web.dto.SignupMemberRequest;
 import io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.SignupUseCase;
-import io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.command.SignupCommand;
+import io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.command.SignupMemberCommand;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -16,15 +16,18 @@ public class SignupController {
 
     private final SignupUseCase signupUseCase;
 
-    @PostMapping("/signup")
-    public ResponseEntity<Long> signup(@RequestBody @Valid SignupRequest request) {
+    @PostMapping("/api/auth/signup/members")
+    public ResponseEntity<Long> signupMember(@RequestBody @Valid final SignupMemberRequest request) {
 
-        final SignupCommand signupCommand = getCommand(request);
-        return ResponseEntity.ok(signupUseCase.signup(signupCommand));
+        final SignupMemberCommand signupMemberCommand = getCommand(request);
+        return ResponseEntity.ok(signupUseCase.signup(signupMemberCommand));
     }
 
-    private static SignupCommand getCommand(final SignupRequest request) {
-        return new SignupCommand(
+
+
+
+    private SignupMemberCommand getCommand(final SignupMemberRequest request) {
+        return new SignupMemberCommand(
                 request.email(),
                 request.nickname(),
                 request.password(),
