+++ src/main/java/io/wisoft/avocadobackendhexagonal/domain/auth/adapter/in/web/dto/SignupMemberRequest.java
@@ -1,8 +1,8 @@
-package io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.command;
+package io.wisoft.avocadobackendhexagonal.domain.auth.adapter.in.web.dto;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record SignupCommand(
+public record SignupMemberRequest(
         @NotBlank String email,
         @NotBlank String nickname,
         @NotBlank String password,
