+++ src/main/java/io/wisoft/avocadobackendhexagonal/domain/auth/application/port/in/SignupUseCase.java
@@ -1,8 +1,8 @@
 package io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in;
 
-import io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.command.SignupCommand;
+import io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.command.SignupMemberCommand;
 
 public interface SignupUseCase {
 
-    Long signup(final SignupCommand request);
+    Long signup(final SignupMemberCommand request);
 }
