+++ src/main/java/io/wisoft/avocadobackendhexagonal/domain/auth/application/port/in/command/SignupMemberCommand.java
@@ -1,8 +1,8 @@
-package io.wisoft.avocadobackendhexagonal.domain.auth.adapter.in.web.dto;
+package io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.command;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record SignupRequest(
+public record SignupMemberCommand(
         @NotBlank String email,
         @NotBlank String nickname,
         @NotBlank String password,
