+++ src/main/java/io/wisoft/avocadobackendhexagonal/domain/auth/application/service/SignUpService.java
@@ -1,7 +1,7 @@
 package io.wisoft.avocadobackendhexagonal.domain.auth.application.service;
 
 import io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.SignupUseCase;
-import io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.command.SignupCommand;
+import io.wisoft.avocadobackendhexagonal.domain.auth.application.port.in.command.SignupMemberCommand;
 import io.wisoft.avocadobackendhexagonal.domain.member.application.port.out.SaveMemberPort;
 import io.wisoft.avocadobackendhexagonal.domain.member.domain.Member;
 import lombok.RequiredArgsConstructor;
@@ -14,7 +14,7 @@ public class SignUpService implements SignupUseCase {
     private final SaveMemberPort saveMemberPort;
 
     @Override
-    public Long signup(final SignupCommand request) {
+    public Long signup(final SignupMemberCommand request) {
 
         final Member member = Member.createMember(
                 null,
